refactor namspace

+++ tests/ui/rpc/fail/async_function_returning_impl_stream.rs
@@ -5,7 +5,7 @@ struct A;
 
 #[rpc]
 impl A {
-    async fn test(&mut self) -> impl futures::Stream<Item = usize> {
+    async fn test(&mut self) -> impl futures::Stream<Item = u8> {
         futures::stream::once(async { 0 })
     }
 }
