+++ tests/ui/rpc/fail/function_returning_impl_iterator.rs
@@ -5,7 +5,7 @@ struct A;
 
 #[rpc]
 impl A {
-    fn test(&mut self) -> impl std::iter::Iterator<Item = usize> {
+    fn test(&mut self) -> impl std::iter::Iterator<Item = u8> {
         std::iter::once(0)
     }
 }
